requirements.txt
pillow ~= 10.3.0
pathlib ~= 1.0.1
